On branch master Your branch is behind 'origin/master' by 1 commit, and can be fast-forwarded.   (use "git pull" to update your local branch)
Changes to be committed:
	modified:   include/connectbutton.hpp
	modified:   include/graphicsview.hpp
	modified:   include/graphicsviewinterface.hpp
	modified:   include/imageplayer.hpp
	modified:   src/graphicscviewinterface.cpp
	modified:   src/graphicsview.cpp
	modified:   src/imageplayer.cpp

+++ include/graphicsview.hpp
@@ -40,7 +40,7 @@ public:
     inline void   setMinZoom(double minZoom) { m_dMinZoom = minZoom; }
     inline void   setMaxZoom(double maxZoom) { m_dMaxZoom = maxZoom; }
     inline QPoint getMousePosition() { return m_qtLastMousePos; }
-    inline void   dynamicMode(ushort _time) { m_pTimer->start(_time); }
+    inline void   dynamicMode(int _time) { m_pTimer->start(_time); }
     inline void   staticMode() { m_pTimer->stop(); }
 
 protected:
